+++ app-bazlang/src/main/java/com/davidconneely/bazlang/MainClass.java
@@ -5,6 +5,7 @@ import com.davidconneely.bazlang.edit.ProgramEditor;
 import com.davidconneely.bazlang.exec.EvalState;
 import com.davidconneely.bazlang.exec.Interpreter;
 import com.davidconneely.bazlang.exec.StatementExecutor;
+import com.davidconneely.bazlang.io.JavaSoundSpeaker;
 import com.davidconneely.bazlang.io.StreamScreen;
 import com.davidconneely.bazlang.io.TerminalScreen;
 import com.davidconneely.bazlang.io.VirtualInput;
@@ -22,8 +23,9 @@ public class MainClass {
   public static void main(String[] args) {
     int exitCode = 0;
     if (System.console() != null) {
-      try (var term = new TerminalScreen(new JLineTerminalEngine())) {
-        exitCode = dispatch(args, term, term, term);
+      try (var term = new TerminalScreen(new JLineTerminalEngine());
+          var speaker = new JavaSoundSpeaker()) {
+        exitCode = dispatch(args, term, term, speaker);
       } catch (IOException ignored) {
       }
     } else {